modules/kvs: Rename fence.[ch] to treq.[ch]
